LegendsToBeP2/backend/requirements.txt
Flask
openai
python-dotenv
requests
selenium
